requirements.txt
discord.py
prsaw
discord-py-slash-command
DiscordUtils
aiosqlite
topggpy
discord_components
youtube_dl
PyNaCl
aiofiles
giphy_client
spotipy
wavelink>=2.0.0